Fix product image migration
The product main image was not being imported because the field name is the new table structure is different from the old name. As a result, all products single pages were displayed without an image.

The postmeta that stores the image ID is called `_thumbnail_id` but the field name in the table `wp_wc_products` is called `image_id`. This commit fixes this issue by simply renaming the field when importing product data from the old table structure.