Remove filtering by term when ordering products by price
Filtering by term when ordering products by price was added for performance reasons due to post meta slowness. On a quick test using the new table structure and a database with 100,000 products I couldn't see any significant difference between the query with and without the term filter. So this commit removes it. We can revisit this later if necessary.