Implement WC_Product_Data_Store_Custom_Table::search_products()
This commit implements the product data store method to search products. It was based on the latest version of the WC core method but adapted to use wp_wc_products when searching for sku and product type (virtual or downloadable).

After the changes made in this commit, the test WC_Tests_Product_Data_Store::test_search_products() is now passing.