Include custom product tables tests in the phpunit.xml file
This commit updates the phpunit.xml file to include custom product tables own tests in the list of tests to run. Before, phpunit was configured to run only WC core tests.

To avoid conflicts between the two different tests suites, it was necessary to change the way WC_Product_Tables_Migrate_Data defines if the data migration is currently running or not. The original method used a constant which couldn't be changed after the migration was over.