Remove download changes
Early in the development of this plugin we decided to store download limit and download expiry per file instead of per product as WC core does. This is a nice change to have but probably something to do first in WC core to make sure the plugin is fully compatible with it. So this commit removes all the changes that were required to implement download limit and download expiry per file and adds two new fields to wp_wc_products to store this information per product.

Related #89